@@ -16,43 +21,66 @@
         "knplabs/knp-menu-bundle": "^2.0",
         "league/glide-symfony": "^1.0",
         "miljar/php-exif": "^0.6.4",
+        "nelmio/cors-bundle": "^1.5",
         "nesbot/carbon": "^1.34",
+        "php-translation/symfony-bundle": "^0.8",
+        "phpdocumentor/reflection-docblock": "^4.2",
         "sensio/framework-extra-bundle": "^5.1",
-        "sensiolabs/security-checker": "^4.1",
+        "sensiolabs/security-checker": "^5.0",
         "siriusphp/upload": "^2.1",
         "stof/doctrine-extensions-bundle": "^1.3",
-        "symfony/asset": "^4.1",
-        "symfony/console": "^4.1",
-        "symfony/debug-pack": "^1.0",
+        "symfony/asset": "^4.2",
+        "symfony/console": "^4.2",
+        "symfony/expression-language": "^4.2",
         "symfony/flex": "^1.1",
-        "symfony/form": "^4.1",
-        "symfony/framework-bundle": "^4.1",
-        "symfony/orm-pack": "^1.0",
+        "symfony/form": "^4.2",
+        "symfony/framework-bundle": "^4.2",
+        "symfony/monolog-bundle": "^3.0",
         "symfony/polyfill-php72": "^1.8",
+        "symfony/security-bundle": "^4.2",
+        "symfony/serializer": "^4.2",
         "symfony/swiftmailer-bundle": "^3.1",
-        "symfony/translation": "^4.1",
-        "symfony/yaml": "^4.1",
+        "symfony/translation": "^4.2",
+        "symfony/twig-bundle": "^4.2",
+        "symfony/validator": "^4.2",
+        "symfony/var-dumper": "^4.2",
+        "symfony/webpack-encore-bundle": "^1.0",
+        "symfony/yaml": "^4.2",
         "tightenco/collect": "^5.7",
         "twig/extensions": "^1.5",
         "webmozart/path-util": "^2.3",
+        "webonyx/graphql-php": "^0.13",
         "white-october/pagerfanta-bundle": "^1.1"
     },
     "require-dev": {
+        "ext-pdo_sqlite": "*",
+        "behat/behat": "^3.5",
+        "behat/mink": "dev-master@dev",
+        "behat/mink-extension": "^2.3",
+        "behat/mink-goutte-driver": "^1.2",
+        "behatch/contexts": "^3.2",
+        "coduo/php-matcher": "dev-feature/array-universal-key@dev",
         "dama/doctrine-test-bundle": "^5.0",
         "doctrine/doctrine-fixtures-bundle": "^3.0",
-        "friendsofphp/php-cs-fixer": "^2.12",
-        "phpstan/phpstan": "^0.10.3",
-        "symfony/browser-kit": "^4.1",
-        "symfony/css-selector": "^4.1",
-        "symfony/debug-bundle": "^4.1",
-        "symfony/dotenv": "^4.1",
-        "symfony/maker-bundle": "^1.7",
-        "symfony/phpunit-bridge": "^4.1",
-        "symfony/stopwatch": "^4.1",
-        "symfony/web-profiler-bundle": "^4.1",
-        "symfony/web-server-bundle": "^4.1"
+        "lakion/mink-debug-extension": "^1.2",
+        "phpspec/phpspec": "^5.1",
+        "phpstan/phpstan": "^0.11",
+        "phpunit/phpunit": "^7.0",
+        "roave/security-advisories": "dev-master@dev",
+        "symfony/browser-kit": "^4.2",
+        "symfony/css-selector": "^4.2",
+        "symfony/debug-bundle": "^4.2",
+        "symfony/dotenv": "^4.2",
+        "symfony/maker-bundle": "^1.11",
+        "symfony/stopwatch": "^4.2",
+        "symfony/web-profiler-bundle": "^4.2",
+        "symfony/web-server-bundle": "^4.2",
+        "symplify/easy-coding-standard": "^5.2"
     },
-    "minimum-stability": "beta",
+    "repositories": [
+        { "type": "vcs", "url": "https://github.com/jarjak/php-matcher", "no-api": true }
+    ],
+    "minimum-stability": "stable",
     "prefer-stable": true,
     "config": {
         "platform": {
